desktop: Moved App struct to app.rs
